+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.0 - ?
+
+* Update `superbalist/php-event-pubsub` to ^2.0
+* Change `pubsub.events` to resolve as a singleton
+* Added support for "Attribute Injectors"
+* Added new `attribute_injectors` config key
+
 ## 1.0.0 - 2017-01-30
 
 * Initial release